execution/requirements.txt
aiohttp==3.10.11
websockets==13.1
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
aiosqlite==0.20.0
